+++ src/atrament/scene-processors.js
@@ -1,0 +1,20 @@
+// Atrament scene processors
+
+export function sceneAddUuid(atrament) {
+  atrament.game.defineSceneProcessor((scene) => {
+    scene.uuid = Date.now();
+  });
+}
+
+export function sceneListImages(atrament) {
+  atrament.game.defineSceneProcessor((scene) => {
+    scene.images = [];
+    scene.content.forEach(item => {
+      if (item.tags.IMAGE) {
+        const imageFullPath = atrament.game.getAssetPath(item.tags.IMAGE);
+        scene.images.push(imageFullPath);
+        item.image = imageFullPath;
+      }
+    });
+  });
+}
