--- src/components/screens/scenes/index.css
@@ -1,19 +1,0 @@
-.block_scene {
-    flex-grow: 1; 
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.block {
-    display: block;
-    padding: 1rem;
-}
-
-.block p:first-child {
-    margin-top: 0px;
-}
-
-.block p:last-child {
-    margin-bottom: 0px;
-}
