+++ src/components/settings/settings-animation/index.js
