+++ src/components/settings/settings-sound/index.css
