+++ src/components/settings/settings-text/index.css
