+++ src/components/ui/container-choices/index.css
@@ -1,3 +1,3 @@
-.block_choices {
+.container_choices {
     flex-grow: 0;
 }
