+++ src/components/ui/container-scenes/index.css
@@ -1,0 +1,6 @@
+.container_scenes {
+    flex-grow: 1; 
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+}
