--- src/components/ui/index.js
@@ -1,41 +1,0 @@
-import ApplicationWrapper from './application-wrapper';
-import Block from './block';
-import ChoiceButton from './choice-button';
-import Container from './container';
-import ContainerFlex from './container-flex';
-import ContainerText from './container-text';
-import Header from './header';
-import ContainerImage from './container-image';
-import LinkHome from './link-home';
-import LinkMenu from './link-menu';
-import SettingsAnimation from './settings-animation';
-import Backdrop from './backdrop';
-import CloseButton from './close-button';
-import Modal from './modal';
-import SettingsFont from './settings-font';
-import SettingsSound from './settings-sound';
-import SettingsText from './settings-text';
-import SettingsTheme from './settings-theme';
-import SettingsVersion from './settings-version';
-
-export {
-  ApplicationWrapper,
-  Block,
-  ChoiceButton,
-  Container,
-  ContainerFlex,
-  ContainerText,
-  Header,
-  ContainerImage,
-  LinkHome,
-  LinkMenu,
-  Backdrop,
-  CloseButton,
-  Modal,
-  SettingsAnimation,
-  SettingsFont,
-  SettingsSound,
-  SettingsText,
-  SettingsTheme,
-  SettingsVersion
-};
