+++ src/components/ui/scene/index.css
@@ -1,0 +1,12 @@
+.scene {
+    display: block;
+    padding: 1rem;
+}
+
+.scene p:first-child {
+    margin-top: 0px;
+}
+
+.scene p:last-child {
+    margin-bottom: 0px;
+}
