+++ src/components/ui/scene/index.js
@@ -1,25 +1,15 @@
 import { h, Fragment } from 'preact';
 import style from './index.css';
-import { useRef, useState, useEffect, useContext, useCallback } from 'preact/hooks';
-import { ContainerImage } from 'src/components/ui';
-import Paragraph from './paragraph';
+import { useRef, useState, useEffect } from 'preact/hooks';
+// UI
+import ContainerImage from '../container-image';
+import Paragraph from '../scene-paragraph';
+// utils
 import preloadImages from 'src/utils/preload-images';
-import Atrament from 'src/atrament-context';
 
 const Scene = ({ scene, isCurrent, readyHandler }) => {
   const [ isLoaded, setIsLoaded ] = useState(false);
-  const atrament = useContext(Atrament);
   const elementRef = useRef(null);
-
-  const getImagesList = useCallback(() => {
-    const images = [];
-    scene.content.forEach(item => {
-      if (item.tags.IMAGE) {
-        images.push(atrament.game.getAssetPath(item.tags.IMAGE));
-      }
-    });
-    return images;
-  }, [ atrament.game, scene.content ]);
 
   useEffect(() => {
     if (isLoaded && isCurrent) {
@@ -34,19 +24,19 @@ const Scene = ({ scene, isCurrent, readyHandler }) => {
   // preload all images for scene
   useEffect(() => {
     const preloader = async () => {
-      await preloadImages(getImagesList());
+      await preloadImages(scene.images);
       setIsLoaded(true);
     }
     preloader();
-  }, [ getImagesList, setIsLoaded ]);
+  }, [ scene, setIsLoaded ]);
 
   return (
-    <div class={[style.block, 'atrament-block-text', isCurrent && isLoaded ? 'animation_appear' : ''].join(' ')} ref={elementRef}>
+    <div class={[style.scene, 'atrament-scene', isCurrent && isLoaded ? 'animation_appear' : ''].join(' ')} ref={elementRef}>
       <div style={{ opacity: isLoaded ? 1 : 0 }}>
         {
           scene.content.map((item, i) => (
             <Fragment key={`paragraph-${scene.uuid}-${i}`}>
-              {item.tags.IMAGE ? <ContainerImage src={atrament.game.getAssetPath(item.tags.IMAGE)} /> : ''}
+              <ContainerImage src={scene.image} />
               <Paragraph isCurrent={isCurrent} content={item.text} />
             </Fragment>
           ))
@@ -57,4 +47,3 @@ const Scene = ({ scene, isCurrent, readyHandler }) => {
 };
 
 export default Scene;
-
